refactor: view update logic to hook
